+++ src/Locator.php
@@ -9,6 +9,7 @@ namespace SimplePie;
 
 use DomDocument;
 use SimplePie\Exception\HttpException;
+use SimplePie\File;
 use SimplePie\HTTP\Client;
 use SimplePie\HTTP\FileClient;
 use SimplePie\HTTP\Response;
@@ -146,7 +147,8 @@ class Locator implements RegistryAware
     public function is_feed(Response $file, bool $check_html = false)
     {
         if (Misc::is_remote_uri($file->get_final_requested_uri())) {
-            $sniffer = $this->registry->create(Content\Type\Sniffer::class, [$file]);
+            $fileResponse = File::fromResponse($file);
+            $sniffer = $this->registry->create(Content\Type\Sniffer::class, [$fileResponse]);
             $sniffed = $sniffer->get_type();
             $mime_types = ['application/rss+xml', 'application/rdf+xml',
                                 'text/rdf', 'application/atom+xml', 'text/xml',